refactor: split decorator logic
